huge overhaul of everything
storing dataset in pickles means it all has to fit in memory, which is no good.
now, use h5py to keep datasets in hdf5 files, which we treat as disk-backed
numpy arrays. convert the entire data set to something like numpy and do cuts
later.

loading events from root is the bottleneck, so use zeromq to distribute it
across a cluster.